Update README.md
Dr. Rochelle Tractenberg is a research methodologist specializing in designs and analyses with "difficult to measure" outcomes in biomedical and educational studies. She is a tenured professor in the Department of Neurology, with secondary appointments in the Departments of Biostatistics, Bioinformatics & Biomathematics and Rehabilitation Medicine at Georgetown University. She is also a Research Fellow at the National Rehabilitation Hospital in Washington, DC. That’s her “day job”; this site is an exploration of what she does “for fun”: innovative stewardship in science, practice, and higher education. You will find the occasional blog post, but the site is more of a sign post to help visitors find support for their stewardship, teaching, learning, statistical practice, or combination-of-these activities. 

Pronouns: she/her/hers

ALL MATERIALS HERE ARE SHARED UNDER cc- by attribution- non commercial – no derivatives 4.0 international license
5 August 2021